Removed checking if console/player ran command - the plugin is non-interactive, neither console or admin with permissions will be different.
Added comments explaining what each block does

Rewrote how the plugin will get the list of regions available. Originally it got the region list of the world the player running the command is in. Now, the plugin will get a full list of all regions across all worlds dynamically and in real-time.